+++ device_frame/lib/device_frame.dart
@@ -3,6 +3,5 @@ library device_frame;
 export 'src/devices.dart';
 export 'src/frame.dart';
 export 'src/info/info.dart';
-export 'src/info/parsing.dart';
 export 'src/keyboard/virtual_keyboard.dart';
 export 'src/theme.dart';