+++ device_frame/lib/src/devices.dart
@@ -1,12 +1,31 @@
-import 'package:device_frame/device_frame.dart';
-import 'package:flutter/foundation.dart';
+import 'dart:ui';
 
+import 'package:device_frame/device_frame.dart';
+import 'package:flutter/widgets.dart';
+import 'package:path_drawing/path_drawing.dart';
+
+import 'info/info.dart';
+
+part 'devices.g.dart';
+
+/// A device form factor.
 enum DeviceType {
+  /// Unknown type
   unknown,
+
+  /// Smartphone
   phone,
+
+  /// Tablet
   tablet,
+
+  /// TV
   tv,
+
+  /// Desktop computer
   desktop,
+
+  /// Laptop
   laptop,
 }
 
@@ -17,18 +36,16 @@ enum DeviceType {
 /// * [loadDeviceInfo] to load information store in assets.
 /// * [DeviceFrame] to display the frame associated to one of identifiers.
 class DeviceIdentifier {
+  /// The unique name of the device (preferably in snake case).
   final String name;
+
+  /// The device form factor.
   final DeviceType type;
+
+  /// The target platform supported by this device.
   final TargetPlatform platform;
 
-  String get assetKey {
-    final platformKey =
-        platform.toString().replaceAll('${TargetPlatform}.', '').toLowerCase();
-    final typeKey =
-        type.toString().replaceAll('${DeviceType}.', '').toLowerCase();
-    return 'packages/device_frame/assets/${platformKey}_${typeKey}_${name}.svg';
-  }
-
+  /// Private constructor.
   const DeviceIdentifier._(
     this.platform,
     this.type,
@@ -49,63 +66,83 @@ class DeviceIdentifier {
   @override
   int get hashCode =>
       runtimeType.hashCode ^ name.hashCode ^ type.hashCode ^ platform.hashCode;
+
+  @override
+  String toString() {
+    final platformKey =
+        platform.toString().replaceAll('${TargetPlatform}.', '').toLowerCase();
+    final typeKey =
+        type.toString().replaceAll('${DeviceType}.', '').toLowerCase();
+    return '${platformKey}_${typeKey}_${name}';
+  }
 }
 
 /// A list of common device specifications sorted by target platform.
 abstract class Devices {
-  static final ios = const IosDevices._();
-  static final android = const AndroidDevices._();
-  static final macos = const MacOsDevices._();
-  static final windows = const WindowsDevices._();
-  static final linux = const LinuxDevices._();
+  static final ios = IosDevices._();
+  static final android = AndroidDevices._();
+  static final macos = MacOsDevices._();
+  static final windows = WindowsDevices._();
+  static final linux = LinuxDevices._();
+
+  static List<DeviceInfo> get all => _allDevices;
 }
 
-class IosDevices {
-  const IosDevices._();
+DeviceInfo _find(TargetPlatform platform, DeviceType type, String name) =>
+    _allDevices.firstWhere(
+      (x) =>
+          x.identifier.platform == platform &&
+          x.identifier.type == type &&
+          x.identifier.name == name,
+    );
 
-  final iPhoneSE = const DeviceIdentifier._(
+/// A set of iOS devices.
+class IosDevices {
+  IosDevices._();
+
+  final iPhoneSE = _find(
     TargetPlatform.iOS,
     DeviceType.phone,
     'iphone-se',
   );
 
-  final iPhone11 = const DeviceIdentifier._(
+  final iPhone11 = _find(
     TargetPlatform.iOS,
     DeviceType.phone,
     'iphone-11',
   );
 
-  final iPhone11Pro = const DeviceIdentifier._(
+  final iPhone11Pro = _find(
     TargetPlatform.iOS,
     DeviceType.phone,
     'iphone-11pro',
   );
 
-  final iPhone11ProMax = const DeviceIdentifier._(
+  final iPhone11ProMax = _find(
     TargetPlatform.iOS,
     DeviceType.phone,
     'iphone-11promax',
   );
 
-  final iPadMini = const DeviceIdentifier._(
+  final iPadMini = _find(
     TargetPlatform.iOS,
     DeviceType.tablet,
     'ipad-mini',
   );
 
-  final iPadPro129 = const DeviceIdentifier._(
+  final iPadPro129 = _find(
     TargetPlatform.iOS,
     DeviceType.tablet,
     'ipad-pro12-9',
   );
 
-  final iPadiPad = const DeviceIdentifier._(
+  final iPadiPad = _find(
     TargetPlatform.iOS,
     DeviceType.tablet,
     'ipad',
   );
 
-  List<DeviceIdentifier> get all => [
+  List<DeviceInfo> get all => [
         iPadMini,
         iPadPro129,
         iPadiPad,
@@ -116,58 +153,59 @@ class IosDevices {
       ];
 }
 
+/// A set of Android devices.
 class AndroidDevices {
-  const AndroidDevices._();
+  AndroidDevices._();
 
-  final small = const DeviceIdentifier._(
+  final small = _find(
     TargetPlatform.android,
     DeviceType.phone,
     'small',
   );
 
-  final medium = const DeviceIdentifier._(
+  final medium = _find(
     TargetPlatform.android,
     DeviceType.phone,
     'medium',
   );
 
-  final large = const DeviceIdentifier._(
+  final large = _find(
     TargetPlatform.android,
     DeviceType.phone,
     'large',
   );
 
-  final pixel3 = const DeviceIdentifier._(
+  final pixel3 = _find(
     TargetPlatform.android,
     DeviceType.phone,
     'pixel3',
   );
 
-  final samsungS8 = const DeviceIdentifier._(
+  final samsungS8 = _find(
     TargetPlatform.android,
     DeviceType.phone,
     'samsung-s8',
   );
 
-  final samsungS20 = const DeviceIdentifier._(
+  final samsungS20 = _find(
     TargetPlatform.android,
     DeviceType.phone,
     'samsung-s20',
   );
 
-  final samsungNote10Plus = const DeviceIdentifier._(
+  final samsungNote10Plus = _find(
     TargetPlatform.android,
     DeviceType.phone,
     'samsung-note10plus',
   );
 
-  final nexus9 = const DeviceIdentifier._(
+  final nexus9 = _find(
     TargetPlatform.android,
     DeviceType.tablet,
     'nexus9',
   );
 
-  List<DeviceIdentifier> get all => [
+  List<DeviceInfo> get all => [
         small,
         medium,
         large,
@@ -179,49 +217,50 @@ class AndroidDevices {
       ];
 }
 
+/// A set of macOS devices.
 class MacOsDevices {
-  const MacOsDevices._();
+  MacOsDevices._();
 
-  final iMacPro = const DeviceIdentifier._(
+  final iMacPro = _find(
     TargetPlatform.macOS,
     DeviceType.desktop,
     'imac-pro',
   );
 
-  final macbook = const DeviceIdentifier._(
+  final macbook = _find(
     TargetPlatform.macOS,
     DeviceType.laptop,
     'macbook',
   );
 
-  List<DeviceIdentifier> get all => [
+  List<DeviceInfo> get all => [
         iMacPro,
         macbook,
       ];
 }
 
 class WindowsDevices {
-  const WindowsDevices._();
+  WindowsDevices._();
 
-  final surface3 = const DeviceIdentifier._(
+  final surface3 = _find(
     TargetPlatform.windows,
     DeviceType.tablet,
     'surface3',
   );
 
-  final screen = const DeviceIdentifier._(
+  final screen = _find(
     TargetPlatform.windows,
     DeviceType.desktop,
     'screen',
   );
 
-  final dellXps13 = const DeviceIdentifier._(
+  final dellXps13 = _find(
     TargetPlatform.windows,
     DeviceType.laptop,
     'dell-xps13',
   );
 
-  List<DeviceIdentifier> get all => [
+  List<DeviceInfo> get all => [
         surface3,
         screen,
         dellXps13,
@@ -229,15 +268,15 @@ class WindowsDevices {
 }
 
 class LinuxDevices {
-  const LinuxDevices._();
+  LinuxDevices._();
 
-  final screen = const DeviceIdentifier._(
+  final screen = _find(
     TargetPlatform.linux,
     DeviceType.desktop,
     'screen',
   );
 
-  List<DeviceIdentifier> get all => [
+  List<DeviceInfo> get all => [
         screen,
       ];
 }